+++ packages/amplify-category-auth/src/provider-utils/awscloudformation/assets/cognito-defaults.js
@@ -32,7 +32,7 @@ const userPoolDefaults = projectName => {
     smsVerificationMessage: 'Your verification code is {####}',
     emailVerificationSubject: 'Your verification code',
     emailVerificationMessage: 'Your verification code is {####}',
-    defaultPasswordPolicy: booleanOptions.find(b => b.value === false).value,
+    defaultPasswordPolicy: false,
     passwordPolicyMinLength: 8,
     passwordPolicyCharacters: [],
     requiredAttributes: ['email'],
@@ -58,7 +58,7 @@ const identityPoolDefaults = projectName => {
   // eslint-disable-line
   return {
     identityPoolName: `${projectName}_identitypool_${sharedId}`,
-    allowUnauthenticatedIdentities: booleanOptions.find(b => b.value === false).value,
+    allowUnauthenticatedIdentities: false,
   };
 };