+++ packages/amplify-category-function/src/commands/function/remove.ts
@@ -1,4 +1,6 @@
 import { category as categoryName } from '../../constants';
+import { ServiceName } from '../../provider-utils/awscloudformation/utils/constants';
+import { removeLayerArtifacts } from '../../provider-utils/awscloudformation/utils/storeResources';
 
 const subcommand = 'remove';
 
@@ -15,6 +17,11 @@ module.exports = {
             'When you delete a layer version, you can no longer configure functions to use it.\nHowever, any function that already uses the layer version continues to have access to it.',
         },
         serviceSuffix: { Lambda: '(function)', LambdaLayer: '(layer)' },
+      })
+      .then((resource: { service: string; resourceName: string }) => {
+        if (resource.service === ServiceName.LambdaLayer) {
+          removeLayerArtifacts(context, resource.resourceName);
+        }
       })
       .catch(err => {
         context.print.info(err.stack);