+++ packages/amplify-cli/src/packageManagerHelpers.ts
@@ -1,6 +1,16 @@
 import * as path from 'path';
 import * as fs from 'fs-extra';
 import * as which from 'which';
+import * as _ from 'lodash';
+import { JSONUtilities } from 'amplify-cli-core';
+
+const packageJson = 'package.json';
+const packageJsonDir = path.join(process.cwd(), packageJson);
+
+const initializationScripts = ['start', 'serve'];
+const MISSING_SCRIPTS_ERROR = new Error(
+  'Did not find a "start" or "serve" initialization script. Add a package.json file in the root of the project with one of these scripts.',
+);
 
 /**
  * Determine the package manager of the current project
@@ -10,8 +20,6 @@ import * as which from 'which';
 export async function getPackageManager() {
   const yarnLock = './yarn.lock';
   const yarnLockDir = path.join(process.cwd(), yarnLock);
-  const packageJson = './package.json';
-  const packageJsonDir = path.join(process.cwd(), packageJson);
   if (fs.existsSync(yarnLockDir)) {
     // Check that yarn is installed for the user
     if (which.sync('yarn', { nothrow: true }) || which.sync('yarn.cmd', { nothrow: true })) {
@@ -22,6 +30,21 @@ export async function getPackageManager() {
     return 'npm';
   }
   return undefined;
+}
+
+/**
+ * Determine the starting command of the current project
+ *
+ * @return {string} 'serve' or 'start'
+ */
+export async function getPackageManagerCommand() {
+  const scripts = _.get(JSONUtilities.readJson(packageJsonDir, { throwIfNotExist: false }), 'scripts');
+  return (
+    _.keys(scripts).find(scriptName => initializationScripts.includes(scriptName)) ||
+    (() => {
+      throw MISSING_SCRIPTS_ERROR;
+    })()
+  );
 }
 
 /**