+++ packages/amplify-e2e-tests/src/environment/env.ts
@@ -1,8 +1,8 @@
 import { nspawn as spawn, getCLIPath, getSocialProviders } from 'amplify-e2e-core';
 
-export function addEnvironment(cwd: string, settings: any) {
+export function addEnvironment(cwd: string, settings: { envName: string; numLayers: number }) {
   return new Promise((resolve, reject) => {
-    spawn(getCLIPath(), ['env', 'add'], { cwd, stripColors: true })
+    const chain = spawn(getCLIPath(), ['env', 'add'], { cwd, stripColors: true })
       .wait('Do you want to use an existing environment?')
       .sendLine('n')
       .wait('Enter a name for the environment')
@@ -10,19 +10,23 @@ export function addEnvironment(cwd: string, settings: any) {
       .wait('Do you want to use an AWS profile?')
       .sendLine('yes')
       .wait('Please choose the profile you want to use')
-      .sendCarriageReturn()
-      .wait('Initialized your environment successfully.')
-      .run((err: Error) => {
-        if (!err) {
-          resolve();
-        } else {
-          reject(err);
-        }
-      });
+      .sendCarriageReturn();
+
+    for (let i = 0; i < settings.numLayers || 0; ++i) {
+      chain.wait('Choose the environment to import the layer access settings from:').sendCarriageReturn(); // Choose first env in list
+    }
+
+    chain.wait('Initialized your environment successfully.').run((err: Error) => {
+      if (!err) {
+        resolve();
+      } else {
+        reject(err);
+      }
+    });
   });
 }
 
-export function checkoutEnvironment(cwd: string, settings: any) {
+export function checkoutEnvironment(cwd: string, settings: { envName: string }) {
   return new Promise((resolve, reject) => {
     spawn(getCLIPath(), ['env', 'checkout', settings.envName], { cwd, stripColors: true })
       .wait('Initialized your environment successfully.')
@@ -37,7 +41,7 @@ export function checkoutEnvironment(cwd: string, settings: any) {
 }
 
 // Test multiple Environments by passing settings.numEnv
-export function listEnvironment(cwd: string, settings: any) {
+export function listEnvironment(cwd: string, settings: { numEnv: number }) {
   return new Promise((resolve, reject) => {
     let numEnv = settings.numEnv || 1;
     let regex = /\|\s\*?[a-z]{2,10}\s+\|/;
@@ -60,7 +64,7 @@ export function listEnvironment(cwd: string, settings: any) {
 }
 
 // Get environment details and return them as JSON
-export function getEnvironment(cwd: string, settings: any): Promise<string> {
+export function getEnvironment(cwd: string, settings: { envName: string }): Promise<string> {
   const envData = {};
   let helper = output => {
     let keyVal = output.split(/:(.+)/); // Split string on first ':' only
@@ -97,7 +101,7 @@ export function getEnvironment(cwd: string, settings: any): Promise<string> {
   but nexpect can't wait() on the spinner output
   See amplify-cli/src/initialize-env.js
 */
-export function pullEnvironment(cwd: string, settings: any) {
+export function pullEnvironment(cwd: string) {
   return new Promise((resolve, reject) => {
     spawn(getCLIPath(), ['env', 'pull'], { cwd, stripColors: true }).run((err: Error) => {
       if (!err) {
@@ -109,7 +113,7 @@ export function pullEnvironment(cwd: string, settings: any) {
   });
 }
 
-export function addEnvironmentHostedUI(cwd: string, settings: any) {
+export function addEnvironmentHostedUI(cwd: string, settings: { envName: string }) {
   const { FACEBOOK_APP_ID, FACEBOOK_APP_SECRET, GOOGLE_APP_ID, GOOGLE_APP_SECRET, AMAZON_APP_ID, AMAZON_APP_SECRET } = getSocialProviders();
   return new Promise((resolve, reject) => {
     spawn(getCLIPath(), ['env', 'add'], { cwd, stripColors: true })
@@ -144,7 +148,7 @@ export function addEnvironmentHostedUI(cwd: string, settings: any) {
   });
 }
 
-export function importEnvironment(cwd: string, settings: any) {
+export function importEnvironment(cwd: string, settings: { envName: string; providerConfig: string }) {
   const cmd_array = [
     'env',
     'import',
@@ -170,7 +174,7 @@ export function importEnvironment(cwd: string, settings: any) {
   });
 }
 
-export function removeEnvironment(cwd: string, settings: any) {
+export function removeEnvironment(cwd: string, settings: { envName: string }) {
   return new Promise((resolve, reject) => {
     spawn(getCLIPath(), ['env', 'remove', settings.envName], { cwd, stripColors: true })
       .wait(`Are you sure you want to continue?`)