+++ packages/graphql-auth-transformer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.20.5](https://github.com/aws-amplify/amplify-cli/compare/graphql-auth-transformer@6.20.4...graphql-auth-transformer@6.20.5) (2020-09-02)
+
+**Note:** Version bump only for package graphql-auth-transformer
+
+
+
+
+
 ## [6.20.4](https://github.com/aws-amplify/amplify-cli/compare/graphql-auth-transformer@6.20.3...graphql-auth-transformer@6.20.4) (2020-08-31)
 
 **Note:** Version bump only for package graphql-auth-transformer